ART: Fix exception stash in stack dump
It is the executing thread that needs its exception stashed. This
may be different from the dumped thread, which, in case it is not
the same thread, the verifier does not care about.

Use RAII for stashing. Make the code simpler by removing the flag
and instead use a non-null exception object as flag.

Bug: 133764636
Test: m test-art-host
Change-Id: I2b6a44f41880c96a0cf9bb343db22963dca0a758
Signed-off-by: Akash <akashniki@gmail.com>